Malicious malloc to use with LD_PRELOAD
This malicious version of malloc
is supposed to allocated buffer to
the caller and also send the buffer
to a RDMA connection. Then remote
attacker can fetch the buffer even
after freeing it.

NOTE: This implementation will not
work as the RDMA library internally
call malloc, so the is a infinite
recursion.